ì 게시글 전체 리스트 나의 리스트 작업 완료

+++ src/main/java/org/example/article/entity/Article.java
@@ -10,25 +10,25 @@ import java.util.Map;
 @Setter
 @AllArgsConstructor
 public class Article {
-    private long id;
+    private int id;
     private String category;
     private String foodName;
     private String brandName;
-    private long price;
-    private long weight;
-    private long scope;
+    private int price;
+    private int weight;
+    private int scope;
     private String review;
     private String writer;
     private String regDate;
 
     public Article (Map<String, Object> row) {
-        this.id = (long)row.get("id");
+        this.id = (int)row.get("id");
         this.category = (String) row.get("category");
         this.foodName = (String) row.get("foodName");
         this.brandName = (String) row.get("brandName");
-        this.price = (long) row.get("price");
-        this.weight = (long) row.get("weight");
-        this.scope = (long) row.get("scope");
+        this.price = (int) row.get("price");
+        this.weight = (int) row.get("weight");
+        this.scope = (int) row.get("scope");
         this.review = (String) row.get("review");
         this.writer = (String) row.get("writer");
         this.regDate = (String)row.get("regDate");
